requirements.txt
# # Core
aiogram==3.26.0
aiohttp==3.13.3
python-dotenv==1.2.2

# Database & State
motor==3.7.1
pymongo==4.16.0
dnspython==2.8.0
redis==7.3.0

# Performance
orjson==3.11.7
aiofiles==25.1.0
aiodns==4.0.0

# Features & Utils
dawg2==0.13.3
matplotlib==3.10.8
psutil==7.2.2
